Created modules and windows directories and moved files to which they belong to.
