+++ src/c/modules/ImageCollection.h
