+++ src/c/modules/TextCollection.c
