+++ src/c/modules/TextCollection.h
